requirements.txt
requests
beautifulsoup4
python-dateutil 